+++ CountlyContentBuilderInternal.h
@@ -25,6 +25,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)changeContent:(NSArray<NSString *> *)tags;
 - (void)refreshContentZone;
 - (void)refreshContentZoneJTE;
+- (void)previewContent:(NSString *)contentId;
 
 #endif
 NS_ASSUME_NONNULL_END