+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## XX.XX.XX
+* Added Content feature method `previewContent:(NSString *)contentId` (Experimental!).
 * Improved refreshing contents.
 
 ## 26.1.0